+++ docs/source/releasenotes.rst
@@ -4,6 +4,10 @@ Release notes
 Upcoming release
 ----------------
 
+- Library **Desktop.Windows**: Add timeout parameter for keywords ``Open From Search```
+  and ``Open Using Run Dialog``
+
+
 6.1.0
 -----